issue 4 - renaming security service and replace model inside a core module
