+++ hexa-interface/src/main/java/com/kitano/iface/KtxEventService.java
