+++ hexa-interface/src/main/java/com/kitano/iface/model/KtxUser.java
