+++ pom.xml
@@ -22,10 +22,11 @@
         <developer/>
     </developers>
     <modules>
-        <module>security-audit-service</module>
         <module>security-notification-service</module>
         <module>security-dashboard</module>
-        <module>core</module>
+        <module>hexa-interface</module>
+        <module>security-core</module>
+        <module>security-service</module>
     </modules>
     <scm>
         <connection/>
