--- security-audit-service/src/main/java/ktx/kitano/audit/service/domain/HomeLabUser.java
@@ -1,48 +1,0 @@
-package ktx.kitano.audit.service.domain;
-
-import com.kitano.iface.model.KtxUser;
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Data
-@NoArgsConstructor
-@Entity
-@Table(name = "users")
-public class HomeLabUser implements KtxUser {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private String id;
-
-    @Column(nullable = false)
-    private String username;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String role;
-
-    @Column(nullable = false)
-    private LocalDateTime created;
-
-    public HomeLabUser(String username, String password, String role) {
-        this.username = username;
-        this.password = password;
-        this.role = role;
-    }
-
-    @PrePersist
-    protected void onCreate() {
-        if (this.id == null) {
-            this.id = UUID.randomUUID().toString();
-        }
-        if (this.created == null) {
-            this.created = LocalDateTime.now();
-        }
-    }
-}
