--- security-audit-service/src/main/java/ktx/kitano/audit/service/infrastructure/repository/DatabaseConfig.java
@@ -1,58 +1,0 @@
-package ktx.kitano.audit.service.infrastructure.repository;
-
-import io.github.cdimascio.dotenv.Dotenv;
-import jakarta.persistence.EntityManagerFactory;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
-import org.springframework.transaction.PlatformTransactionManager;
-
-import javax.sql.DataSource;
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Configuration class for the database
- * Get the database credentials from the .env file
- * Manage the connection to the database
- */
-@Configuration
-public class DatabaseConfig {
-
-    Dotenv dotenv = Dotenv.load();
-
-    String username = dotenv.get("DB_USERNAME");
-    String password = dotenv.get("DB_PASSWORD");
-
-    @Bean
-    public DataSource dataSource() {
-        return DataSourceBuilder.create()
-                .url("jdbc:mysql://localhost:3306/homelab_db")
-                .username(username)
-                .password(password)
-                .driverClassName("com.mysql.cj.jdbc.Driver")
-                .build();
-    }
-
-    @Bean
-    public LocalContainerEntityManagerFactoryBean entityManagerFactory(DataSource dataSource) {
-        LocalContainerEntityManagerFactoryBean em = new LocalContainerEntityManagerFactoryBean();
-        em.setDataSource(dataSource);
-        em.setPackagesToScan("ktx.kitano.audit.service.domain");
-        em.setJpaVendorAdapter(new HibernateJpaVendorAdapter());
-
-        Map<String, Object> properties = new HashMap<>();
-        properties.put("hibernate.dialect", "org.hibernate.dialect.MySQLDialect");
-        em.setJpaPropertyMap(properties);
-
-        return em;
-    }
-
-    @Bean
-    public PlatformTransactionManager transactionManager(EntityManagerFactory entityManagerFactory) {
-        return new JpaTransactionManager(entityManagerFactory);
-    }
-}
