+++ security-dashboard/src/main/java/ktx/kitano/security/service/Main.java
@@ -1,0 +1,7 @@
+package ktx.kitano.security.service;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
