--- security-notification-service/src/main/java/ktx/kitano/audit/service/Main.java
@@ -1,7 +1,0 @@
-package ktx.kitano.audit.service;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
