+++ security-service/src/main/java/ktx/kitano/security/service/infrastructure/messaging/KafkaProducerConfig.java
@@ -1,0 +1,41 @@
+package ktx.kitano.security.service.infrastructure.messaging;
+
+import com.kitano.core.model.SystemEvent;
+import org.apache.kafka.clients.producer.ProducerConfig;
+import org.apache.kafka.common.serialization.StringSerializer;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.core.DefaultKafkaProducerFactory;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.core.ProducerFactory;
+import org.springframework.kafka.support.serializer.JsonSerializer;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Configuration
+public class KafkaProducerConfig {
+
+    @Value(value = "${kafka.bootstrapAddress:kafka:9092}")
+    private String bootstrapAddress;
+
+    @Bean
+    public ProducerFactory<String, SystemEvent> senderFactory() {
+        Map<String, Object> configProps = new HashMap<>();
+        configProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapAddress);
+        configProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
+        configProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
+
+        // Disable type headers to prevent serialization issues
+        JsonSerializer<SystemEvent> jsonSerializer = new JsonSerializer<>();
+        jsonSerializer.setAddTypeInfo(false);
+
+        return new DefaultKafkaProducerFactory<>(configProps, new StringSerializer(), jsonSerializer);
+    }
+
+    @Bean
+    public KafkaTemplate<String, SystemEvent> kafkaTemplate() {
+        return new KafkaTemplate<>(senderFactory());
+    }
+}
