+++ security-service/src/main/java/ktx/kitano/security/service/infrastructure/repository/SecurityEventStore.java
@@ -1,0 +1,12 @@
+package ktx.kitano.security.service.infrastructure.repository;
+
+import com.kitano.iface.KtxEventService;
+import com.kitano.iface.model.KtxEvent;
+
+import java.util.List;
+
+public interface SecurityEventStore<T extends KtxEvent<?>> extends KtxEventService<T> {
+    List<T> findByType(KtxEvent.EventType eventType);
+
+    List<T> findByUserId(String userId);
+}
