+++ security-service/src/main/resources/application.properties
@@ -1,0 +1,44 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/audit_db
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.profiles.active=main-profile
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.id.new_generator_mappings=false
+spring.main.allow-bean-definition-overriding=true
+
+# Enable logging to a file
+logging.file.name=logs/security-service.log
+logging.file.path=logs
+logging.config=classpath:logback.xml
+logging.level.root=INFO
+logging.level.org.springframework=INFO
+logging.level.org.hibernate=INFO
+logging.level.org.apache.kafka=INFO
+
+# Kafka Configuration
+spring.kafka.bootstrap-servers=localhost:9092
+
+# Kafka Producer Configuration
+spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
+spring.kafka.producer.value-serializer=org.springframework.kafka.support.serializer.JsonSerializer
+spring.kafka.consumer.properties.spring.json.value.default.type=com.kitano.core.model.SystemEvent
+spring.kafka.consumer.properties.spring.json.trusted.packages=com.kitano.core.model
+
+# Producer Configuration
+spring.kafka.producer.retries=3
+spring.kafka.producer.acks=all
+spring.kafka.producer.properties.max.block.ms=1000
+
+# Consumer Configuration
+spring.kafka.consumer.group-id=security-group
+spring.kafka.consumer.auto-offset-reset=earliest
+spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.consumer.value-deserializer=org.springframework.kafka.support.serializer.JsonDeserializer
+
+# General Kafka Properties
+spring.kafka.properties.request.timeout.ms=5000
+spring.kafka.properties.reconnect.backoff.ms=1000
